Removed event callback declarations
This duct tape of a solution didn't stick well and caused more problems than it was solving.
It's easier for those using TypeScript to just ensure they use the correct type in the callbacks.
I will try to improve the documentation in such a way that they know which types to use.